rename PHP80 duplications to unique test class names
