+++ packages/coding-standard/tests/Fixer/ArrayNotation/StandaloneLineInMultilineArrayFixer/StandaloneLineInMultilineArrayFixerTestPhp80Test.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Symplify\CodingStandard\Tests\Fixer\LineLength\LineLengthFixer;
+namespace Symplify\CodingStandard\Tests\Fixer\ArrayNotation\StandaloneLineInMultilineArrayFixer;
 
 use Iterator;
 use Symplify\EasyCodingStandard\Testing\PHPUnit\AbstractCheckerTestCase;
 use Symplify\EasyTesting\DataProvider\StaticFixtureFinder;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
-final class Php80Test extends AbstractCheckerTestCase
+final class StandaloneLineInMultilineArrayFixerTestPhp80Test extends AbstractCheckerTestCase
 {
     /**
-     * @dataProvider provideDataForTest()
+     * @dataProvider provideData()
      */
     public function test(SmartFileInfo $fileInfo): void
     {
@@ -20,9 +20,9 @@ final class Php80Test extends AbstractCheckerTestCase
     }
 
     /**
-     * @return Iterator<SmartFileInfo>
+     * @return Iterator<mixed, SmartFileInfo>
      */
-    public function provideDataForTest(): Iterator
+    public function provideData(): Iterator
     {
         return StaticFixtureFinder::yieldDirectoryExclusively(__DIR__ . '/FixturePhp80');
     }
