+++ packages/phpstan-rules/tests/Rules/NoDynamicNameRule/NoDynamicNameRulePhp80Test.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Symplify\PHPStanRules\Tests\Rules\NoParentMethodCallOnEmptyStatementInParentMethodRule;
+namespace Symplify\PHPStanRules\Tests\Rules\NoDynamicNameRule;
 
 use Iterator;
 use PHPStan\Rules\Rule;
 use Symplify\PHPStanExtensions\Testing\AbstractServiceAwareRuleTestCase;
-use Symplify\PHPStanRules\Rules\NoParentMethodCallOnEmptyStatementInParentMethodRule;
+use Symplify\PHPStanRules\Rules\NoDynamicNameRule;
 
 /**
- * @extends AbstractServiceAwareRuleTestCase<NoParentMethodCallOnEmptyStatementInParentMethodRule>
+ * @extends AbstractServiceAwareRuleTestCase<NoDynamicNameRule>
  */
-final class Php80Test extends AbstractServiceAwareRuleTestCase
+final class NoDynamicNameRulePhp80Test extends AbstractServiceAwareRuleTestCase
 {
     /**
      * @dataProvider provideData()
@@ -28,14 +28,11 @@ final class Php80Test extends AbstractServiceAwareRuleTestCase
      */
     public function provideData(): Iterator
     {
-        yield [__DIR__ . '/FixturePhp80/SkipPromotedParentProperty.php', []];
+        yield [__DIR__ . '/FixturePhp8/SkipObjectClassOnPhp8.php', []];
     }
 
     protected function getRule(): Rule
     {
-        return $this->getRuleFromConfig(
-            NoParentMethodCallOnEmptyStatementInParentMethodRule::class,
-            __DIR__ . '/config/configured_rule.neon'
-        );
+        return $this->getRuleFromConfig(NoDynamicNameRule::class, __DIR__ . '/config/configured_rule.neon');
     }
 }
